# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Kafka Integration API Documentation",
        "description": "API documentation for Laravel Kafka Integration project, including table ordering system",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "http://localhost:9999"
        }
    ],
    "paths": {
        "/api/kafka/produce": {
            "post": {
                "tags": [
                    "Kafka"
                ],
                "summary": "Send a message to a Kafka topic",
                "description": "Sends a message to the specified Kafka topic with an optional key",
                "operationId": "produceMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "topic",
                                    "message"
                                ],
                                "properties": {
                                    "topic": {
                                        "type": "string",
                                        "example": "test-topic"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Hello Kafka from Laravel!"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "test-key"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Message sent to Kafka"
                                        },
                                        "data": {
                                            "properties": {
                                                "topic": {
                                                    "type": "string",
                                                    "example": "test-topic"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Hello Kafka from Laravel!"
                                                },
                                                "key": {
                                                    "type": "string",
                                                    "example": "test-key"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}